backend/requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
Werkzeug==3.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
